+++ src/main/java/com/playtika/qa/carsshop/web/CarController.java
@@ -32,7 +32,7 @@ public class CarController {
     }
 
     @GetMapping(value = "/cars/{id}")
-    public CarInfo getCar(@PathVariable(value = "id") Integer id) throws NotFoundException {
+    public CarInfo getCar(@PathVariable(value = "id") long id) throws NotFoundException {
         log.info("get request with id {} received", id);
 
         return service.get(id)
@@ -47,7 +47,7 @@ public class CarController {
     }
 
     @DeleteMapping("cars/{id}")
-    public void deleteCars(@PathVariable("id") Integer id) {
+    public void deleteCars(@PathVariable("id") long id) {
         service.delete(id);
     }
 }
